requirements.txt
streamlit
openai
requests
huggingface-hub
groq
python-dotenv
